Move send_email into its own module. It works!
